refactored ExpandedEventView to use two components: EditEventDialog and DroppedEventDialog and implemented edit functionality on right click context menu
