+++ client/src/components/timetable/EditEventDialog.tsx
@@ -1,0 +1,259 @@
+import React, { useContext, useState } from 'react';
+import { AccessTime, Close, Delete, Edit, Event, LocationOn, Notes, Save } from '@mui/icons-material';
+import { Dialog, Grid, IconButton, ListItem, ListItemIcon, TextField, Typography } from '@mui/material';
+import { TimePicker } from '@mui/x-date-pickers';
+import { daysLong, daysShort } from '../../constants/timetable';
+import { AppContext } from '../../context/AppContext';
+import { CourseContext } from '../../context/CourseContext';
+import { EventTime } from '../../interfaces/Periods';
+import { EditEventDialogProps } from '../../interfaces/PropTypes';
+import { ExecuteButton, StyledListItem, StyledListItemText } from '../../styles/CustomEventStyles';
+import { StyledDialogContent, StyledDialogTitle, StyledTitleContainer } from '../../styles/ExpandedViewStyles';
+import { to24Hour } from '../../utils/convertTo24Hour';
+import { createNewEvent } from '../../utils/createEvent';
+import { useEventDrag } from '../../utils/Drag';
+import { areValidEventTimes, createDateWithTime } from '../../utils/eventTimes';
+import ColorPicker from '../controls/ColorPicker';
+import DiscardDialog from './DiscardDialog';
+import DropdownOption from './DropdownOption';
+import DroppedEventDialog from './DroppedEventDialog';
+
+const EditEventDialog: React.FC<EditEventDialogProps> = ({ eventPeriod, handleCloseDialog, setIsEditing, isChanged, setIsChanged }) => {
+  const { name, location, description, color } = eventPeriod.event;
+  const { day, start, end } = eventPeriod.time;
+
+  const [newName, setNewName] = useState<string>(name);
+  const [newDays, setNewDays] = useState<Array<string>>([daysShort[day - 1]]);
+  const [newStartTime, setNewStartTime] = useState<Date>(createDateWithTime(start));
+  const [newEndTime, setNewEndTime] = useState<Date>(createDateWithTime(end));
+  const [newLocation, setNewLocation] = useState<string>(location);
+  const [newDescription, setNewDescription] = useState<string>(description);
+
+  const [colorPickerAnchorEl, setColorPickerAnchorEl] = useState<HTMLElement | null>(null);
+  const [newColor, setNewColor] = useState<string>(color as string);
+
+  const { createdEvents, setCreatedEvents } = useContext(CourseContext);
+  const { setErrorVisibility, setAlertMsg } = useContext(AppContext);
+
+  const handleOpenColorPicker = (event: React.MouseEvent<HTMLElement>) => {
+    setColorPickerAnchorEl(event.currentTarget);
+  };
+
+  const handleCloseColorPicker = () => {
+    setColorPickerAnchorEl(null);
+  };
+
+  const handleFormat = (newFormats: string[]) => {
+    setNewDays(newFormats);
+    setIsChanged(true);
+  };
+
+  /**
+   * @param isChanged Indicates if an edit has been made to the start of the event
+   * @param newStartTime The new edited start time
+   * @param start The original starting time of the event
+   * @returns The correct time to display on the time picker.
+   * The newStartTime is only displayed if an edit has been made, otherwise the
+   * original start time is shown.
+   */
+  const timePickerStart = (isChanged: boolean, newStartTime: Date, start: number) => {
+    return isChanged ? newStartTime : createDateWithTime(start);
+  };
+
+  /**
+   * @param isChanged Indicates if an edit has been made to the end of the event
+   * @param newEndTime The new edited start time
+   * @param end The original ending time of the event
+   * @returns The correct time to display on the time picker.
+   * The newEndTime is only displayed if an edit has been made, otherwise the
+   * original end time is shown.
+   */
+  const timePickerEnd = (isChanged: boolean, newEndTime: Date, end: number) => {
+    return isChanged ? newEndTime : createDateWithTime(end);
+  };
+
+  const updateEventTime = (eventTime: EventTime, id: string) => {
+    setCreatedEvents({
+      ...createdEvents,
+      [id]: {
+        ...createdEvents[id],
+        time: { ...eventTime },
+      },
+    });
+
+    // Update the time that appears in the TimePicker boxes when in edit mode.
+    setNewDays([daysShort[eventTime.day - 1]]);
+
+    setNewStartTime(createDateWithTime(eventTime.start));
+    setNewEndTime(createDateWithTime(eventTime.end));
+  };
+
+  useEventDrag(updateEventTime);
+
+  const handleUpdateEvent = (id: string) => {
+    if (!areValidEventTimes(newStartTime, newEndTime)) {
+      setAlertMsg('End time is earlier than start time');
+      setErrorVisibility(true);
+      return;
+    }
+
+    // For cloning events
+    if (newDays.length > 1) {
+      createEvents(id);
+      return;
+    }
+
+    const newEventTime = {
+      day: daysShort.indexOf(newDays.toString()) + 1,
+      start: newStartTime.getHours() + newStartTime.getMinutes() / 60,
+      end: newEndTime.getHours() + newEndTime.getMinutes() / 60,
+    };
+
+    setCreatedEvents({
+      ...createdEvents,
+      [id]: {
+        type: 'event',
+        event: {
+          id: id,
+          name: newName,
+          location: newLocation,
+          description: newDescription,
+          color: newColor,
+        },
+        time: newEventTime,
+      },
+    });
+
+    setIsEditing(false);
+  };
+
+  const createEvents = (id: string) => {
+    const updatedEventData = { ...createdEvents };
+    // Delete the original event
+    delete updatedEventData[id];
+
+    // Create an event for each day that is selected in the dropdown option
+    for (const day of newDays) {
+      const newEvent = createNewEvent(newName, newLocation, newDescription, newColor, day, newStartTime, newEndTime);
+      updatedEventData[newEvent.event.id] = newEvent;
+    }
+
+    setCreatedEvents(updatedEventData);
+    setIsEditing(false);
+  };
+
+  return (
+    <>
+        <StyledDialogTitle>
+        <StyledTitleContainer>
+            <Grid container justifyContent="flex-end" alignItems="center">
+            <IconButton aria-label="close" onClick={handleCloseDialog}>
+                <Close />
+            </IconButton>
+            </Grid>
+        </StyledTitleContainer>
+        </StyledDialogTitle>
+        <StyledDialogContent>
+        <ListItem>
+            <ListItemIcon>
+            <Event />
+            </ListItemIcon>
+            <TextField
+            fullWidth={true}
+            id="outlined-required"
+            required
+            variant="outlined"
+            value={newName}
+            onChange={(e) => {
+                setIsChanged(true);
+                setNewName(e.target.value);
+            }}
+            />
+        </ListItem>
+        <ListItem>
+            <ListItemIcon>
+            <Notes />
+            </ListItemIcon>
+            <TextField
+            fullWidth={true}
+            label="Description (optional)"
+            id="outlined-required"
+            variant="outlined"
+            value={newDescription}
+            multiline
+            onChange={(e) => {
+                setIsChanged(true);
+                setNewDescription(e.target.value);
+            }}
+            />
+        </ListItem>
+        <ListItem>
+            <ListItemIcon>
+            <LocationOn />
+            </ListItemIcon>
+            <TextField
+            fullWidth={true}
+            id="outlined-required"
+            required
+            variant="outlined"
+            value={newLocation}
+            onChange={(e) => {
+                setIsChanged(true);
+                setNewLocation(e.target.value);
+            }}
+            />
+        </ListItem>
+        <ListItem>
+            <StyledListItemText primary="Start time" />
+            <TimePicker
+            value={timePickerStart(isChanged, newStartTime, start)}
+            onChange={(e) => {
+                setIsChanged(true);
+                if (e) setNewStartTime(e);
+            }}
+            />
+        </ListItem>
+        <ListItem>
+            <StyledListItemText primary="End time" />
+            <TimePicker
+            value={timePickerEnd(isChanged, newEndTime, end)}
+            label={!areValidEventTimes(newStartTime, newEndTime) ? 'End time must be after start' : ''}
+            slotProps={{ textField: { color: areValidEventTimes(newStartTime, newEndTime) ? 'primary' : 'error' } }}
+            onChange={(e) => {
+                setIsChanged(true);
+                if (e) setNewEndTime(e);
+            }}
+            />
+        </ListItem>
+        <ListItem disablePadding={true}>
+            <DropdownOption
+            optionName="Days"
+            optionState={newDays}
+            setOptionState={handleFormat}
+            optionChoices={daysShort}
+            multiple={true}
+            noOff
+            />
+        </ListItem>
+        <ColorPicker
+            color={newColor}
+            setColor={setNewColor}
+            colorPickerAnchorEl={colorPickerAnchorEl}
+            handleOpenColorPicker={handleOpenColorPicker}
+            handleCloseColorPicker={handleCloseColorPicker}
+        />
+        </StyledDialogContent>
+        <ExecuteButton
+        variant="contained"
+        color="primary"
+        onClick={() => handleUpdateEvent(eventPeriod.event.id)}
+        disabled={newName === '' || newLocation === '' || newDays.length === 0}
+        >
+        <Save />
+        SAVE
+        </ExecuteButton>
+    </>
+  );
+};
+
+export default EditEventDialog;
